pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dungeon-crawl"
version = "2.0.0"
description = "A terminal-based dungeon crawler RPG with roguelike elements"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "stb5585"}
]
keywords = ["game", "rpg", "dungeon-crawler", "roguelike", "terminal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Games/Entertainment :: Role-Playing",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=2.2.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "pytest-cov>=4.1.0",
]
gui = [
    "pygame>=2.5.0",  # For future Phase 3 GUI
    "Pillow>=10.0.0",  # For perspective transforms in dungeon renderer
]

[project.urls]
Homepage = "https://github.com/stb5585/Dungeon-Crawl"
Repository = "https://github.com/stb5585/Dungeon-Crawl"
Issues = "https://github.com/stb5585/Dungeon-Crawl/issues"

[project.scripts]
dungeon-crawl = "game:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["combat*", "effects*", "events*", "presentation*", "analytics*", "data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--tb=short",
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    ".venv/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
